[Feature] Support typeof function (#42627)
The typeof function is completely implemented in fe.
In ExpressionAnalyzer.visitFunctionCall, we can get the parameter type of typeof. At this time, we have already got the result of the typeof function. We only need to directly replace the current FunctionCallExpr with StringLiteral, but I can't get the parent node of typeof, so I can't make this replacement. To solve this problem, I put StringLiteral in the position of the typeof parameter, and then rewrite it in the FoldConstantsRule.

Signed-off-by: changxin <streakxin@foxmail.com>