[BugFix] Fix use after free when read the GIN
Why I'm doing:
This problem is cause by uncorrect usage of the Clucene lib.
The bug can be repeated as following:

1. Use Gin to filtering predicate '!=keyword', in this case, we will call
FullTextCLuceneInvertedReader::query. When this function is finished, it will
call _CLDECDELETE(directory) to destory the dir struct. But this marco actually
do nothing for the directory at all. The refCount of this directory is still 1 and
this object was still stored in a global map managed by Clucene.

2. Use Gin to filtering predicate 'IS NOT NULL', in this case, we will call
FullTextCLuceneInvertedReader::query_null. When this function is finished,
it will CLOSE_INPUT(dir) to destory the dir struct. In this case, the dir
refCounter is 2 and it can not be drop from the global map. But the problem is
that, CLOSE_INPUT will still delete the dir pointer. This makes that the
pointer save in the global map become the dangling pointer.

3. Use Gin to filtering predicate '!=keyword' again, when we want to new a directory
, we first find in the global map, but the element in the map is dangling pointer and
we hit the use after free bug and crash in asan mode.

What I'm doing:
1. Do not delete pointer manually after close. Because FSDirectory::close will delete itself
automatically if the refCount is 0.
2. Use CLOSE_DIR instead of _CLDECDELETE in FullTextCLuceneInvertedReader::query.

Signed-off-by: srlch <linzichao@starrocks.com>