[BugFix] Conflict between Restore and migration cause by storage medium
Why I'm doing:
In currently implementation, restore will always create the new tablet for
the table as HDD storage medium. But the meta data snapshot for storage medium
can be SSD, which cause the problem. Tablet report find that the different medium
between BE side and FE partition metadata, migration will be triggered.
But in the same time, restore is processing, which cause the restore failed.

What I'm doing:
Do not check the storage medium in tablet report for the Restoring tablet.

Signed-off-by: srlch <linzichao@starrocks.com>