[BugFix] Fix the bug that causes incorrect pindex data when the key length is very long (#43568)
## Why I'm doing:
Pindex maybe return error result if the key is very long. We will calculate a hash value for each key, and use this hash value to determine which shard, page, and bucket the key will be distributed to.  And we will record the bucket info (include pageid, packid and size) in the `BucketHeader`. However, the `pageid` in BucketHeader is store as `uint16_t` and the max page num should be not greater than 65536. But now it's possible that the number of pages exceeds 65536. A typical example is following:
1. Assuming each key has a length of 3000 characters.
2. The page size is 4096 right now and each page can only keep one key right now.
3. We determine shard num by the total bytes of all the keys and we will try to generate a immutable pindex.
4. Since we use hashing to determine which page a key should be placed in, it's possible that different keys will be placed in the same page. If that, we will increase the page num and retry until each page only has on key at most. At this time, the page num maybe increase very fast and exceeds 65536.

## What I'm doing:
IMO, A 4k-sized page isn't suitable for storing very long keys, as this leaves little tolerance for hash collisions, resulting in many empty pages and causing performance degradation and storage space waste and When the key length is not very long, the likelihood of exceeding the limit of the number of pages is very small.  So, I supported larger pages to ensure that each page can store at least `kRecordPerBucket` (default value is 8) records to tolerate hash collisions. When the number of pages exceeds the limit, an error will be returned.

Signed-off-by: zhangqiang <qiangzh95@gmail.com>
(cherry picked from commit 9f9d6dc5a7480910a60a0e9d7ec9a4ebe194660d)

# Conflicts:
#	be/src/storage/persistent_index.cpp