[Enhancement] HashJoin no longer output columns that the upper node does not need (#44355)
Why I'm doing:
Before the (branch-2.2) pr (#3042), Hash join node should output all the columns of left and right table, because he don't know which column will be used by upper project node.

After the (branch-2.2) pr (#3042), we will populate a default column for upgrade compatibility.

In branch-3.3 we can remove this logic.

The purpose of doing this is to prepare for the later integration of Join Lazymaterized code.

What I'm doing:
No need to populate a default column for upgrade compatibility.

Signed-off-by: trueeyu <lxhhust350@qq.com>