[BugFix] Fix incorrect column read for auto-increment column in partial update
Why I'm doing:
When we execute partial update for auto-increment column in pk table, we need
to read the partial segment file to get the new allocated id to the segment rewrite.
But we wrongly use the auto-increment column id in full schema to indicate the
column in partial segment, which cause be crash.

What I'm doing:
We should use auto-increment column id in partial segment schema.

Signed-off-by: srlch <linzichao@starrocks.com>