[BugFix] Fix uncorrect unmodified_column_ids if only partial update AUTO_INCREMENT column
Why I'm doing:
There is a very special scene which we only partial update the auto increment column itself
but not other columns. In this case, txn_meta.partial_update_column_unique_ids(), this is
because even we want to partial update AUTO_INCREMENT column, the column will be added into
partial segment. So, in this case, we return unmodified_column_ids as full schema in
get_read_columns_ids which is wrong.

What I'm doing:
Correct the unmodified_column_ids if there is a auto increment column in the schema
if we only partial update AUTO_INCREMENT column.

Signed-off-by: srlch <linzichao@starrocks.com>