[Enhancement] pass connect context to the prepareMetadata and distributed_plan interface (#45404)
Why I'm doing:
we use ConnectContext.get() to get Session variable in the iceberg job planning. If we don't execute planning in the query thread, and can't get the connect context. There are two cases:

In the prepareMetadata, the job planning is executed in the thread pool
distributed plan mode, the metadata collect job is executed in the another thread.
1 && 2. plan_mode is distributed and running under the thread pool
So we need to adapt to the two case with the connect context.

What I'm doing:
Adapting ConnectContext and Tracers under multi-threading

Fixes #issue
#43460

Signed-off-by: stephen <stephen5217@163.com>