[Refactor] Introduce ColumnId to support Column renaming (part1) (#45364)
## Why I'm doing:
We need a unique ID to identify the Column. This ID is used in all places that reference the Column. In this way, to change the attributes of the Column, such as name, we only need to change the attributes in the Column object. There is a uniqueId id the current Column, but it is only available in newly created tables. For compatibility reasons, we need to introduce another Id: ColumnID. The ColumnID of the historical table is the name of the column, because the name was previously immutable.

## What I'm doing:
Add ColumnId object.

Signed-off-by: gengjun-git <gengjun@starrocks.com>
(cherry picked from commit 793c9fec75b819c98a55a6992a47f53819574f9c)

# Conflicts:
#	test/sql/test_column_rename/R/test_column_rename2