[BugFix] Fix query a view with MATCH got exception
Background:
In this pr #44514, we disbale the global dict rewrite
for the left child of the MATCH expression. We record
the left child operator Id of the MATCH expression,
and skip the rewrite if we meet this id again.

But the problem is that, if we query a view with
MATCH expression, the left child of the MATCH expression
may not represent a physical column(which should be
appear in PhysicalOlapScan). And we will still try to rewrite
the corrponding column and got exception.

Solution:
Check all matchChildren id before the rewrite. If there is a
child id is not contained in the physical columns set in PhysicalOlapScan,
it mean that the left child of match expression may be a non-column Ref
expression and we should stop the rewrite immediately.

Signed-off-by: srlch <linzichao@starrocks.com>