[BugFix] Fix JniScanner crash due to struct null indicator
"appendStruct" method does not update the size of the NULLS indicator.
In some cases, there will be problems. For example, when Structs as
Array's children, ARRAY will calculate 4096 in a batch, and Structs
will be more than 4096. If these None of the null in the struct, the
size of the Nulls indicator will never be updated, which will cause
the NULLS indicator size to not match the number of rows of columns.
When passed to BE, the size of the memory copy will be determined
according to the number of rows column. This will cause memory to
cross the border and Crash.

Signed-off-by: changxin <streakxin@foxmail.com>