[BugFix] Fix BE graceful exit use-after-free
Since check_operator_close_states calls Operator::get_name related operations, this calls Operator::is_finished and other related functions. At this point the operator may be in an 'illegal' state (not initialized properly). This will cause a BE crash.
This PR will not prevent the BE from exiting. But it will prevent the BE from generating a SIGSEV when it exits, which could affect the troubleshooting of the problem.

Signed-off-by: stdpain <drfeng08@gmail.com>