[BugFix] Fix stuck in cancel alter table if create tablet take too long time (#47312)
## Why I'm doing:
We may stuck in cancel alter table if it is executing create replica task. Because we wait for the condition variables until all tasks are finished or timeout happen. During the job waiting on the condition variable, cancel alter table will stuck because runPendingJob() will hold the object lock.

## What I'm doing:
Countdown to zero for the condition variables if we want to cancel the job.

Signed-off-by: srlch <linzichao@starrocks.com>