[Doc] change search ranking for partition (#47713)
The current search results for "partition" or "fenqu" are probably not what the reader is looking for. We recommend expression partitioning, but "fenqu" leads to list partitioning. I think that the main data distribution page is the best place.

@EsoragotoSpirit can you fix the description line for the Chinese data distribution page (I put in a placeholder). Once that line is good take off the draft.

At some point I need to fix the way our crawler processes the keywords array, I only tested with single entries, which work fine, but when we add multiple keywords the ranking is not affected. I need to add some Javascript to process the array in the crawler. For now when we need multiple keywords for Pinyin I would like to test using the description field. So, for example I want the words "partition" and "bucket" in the description for data distribution, so I used the phrase "Partition and bucket data". Note that the description should be a short sentence as it appears on overview pages and in search results.

Signed-off-by: DanRoscigno <dan@roscigno.com>
Signed-off-by: 絵空事スピリット <wanglichen@starrocks.com>
Co-authored-by: 絵空事スピリット <wanglichen@starrocks.com>
(cherry picked from commit 703a58898a405f46e2b41de9fa0f6f6bebc5f031)

# Conflicts:
#	docs/en/table_design/Data_distribution.md
#	docs/en/table_design/expression_partitioning.md
#	docs/zh/table_design/Data_distribution.md
#	docs/zh/table_design/list_partitioning.md