[BugFix] Fix forget erase the partition root directory for UnPartitioned table in share data mode
Why I'm doing:
If the we execute INSERT INTO OVERWRITE or OPTIMIZE TABLE for UnPartitioned table, the files for
the old partition will be leaked and can not be removed. The reason is that,
OlapTable/LakeTable::replacePartition interface will not call OlapTable/LakeTable::dropParition
for the old one. And no parition CatalogRecycleBin background task will be generated for the old partition.
But the cleaning job for removing the old partition directory is depending on the background task in
CatalogRecycleBin and we will leak the old partition files in remote storage.

What I'm doing:
Use OlapTable/LakeTable::dropParition instead.

Signed-off-by: srlch <linzichao@starrocks.com>