[BugFix] S3 bytes range contain first and last byte (#49692)
Why I'm doing:
https://www.rfc-editor.org/rfc/rfc9110.html#name-range
The first-pos value in a bytes int-range gives the offset of the first byte in a range. The last-pos value gives the offset of the last byte in the range; that is, the byte positions specified are inclusive. Byte offsets start at zero.

What I'm doing:
s3 byte range contain first and last byte, so the last byte should be _offset + count - 1

Signed-off-by: zombee0 <ewang2027@gmail.com>