[BugFix] Fix pk get error ouput rowset schema during compaction (#49444)
## Why I'm doing:
We will support compaction a part of segments in the future pr and we need to keep the schema consistence among segments in a rowset. So we need to pick an output rowset schema during compaction but not use latest tablet schema instead. 

The last input rowset is always the latest rowset if table is non-primary key table. However, if the table is primary key table, the last input rowset maybe generated by compaction and maybe not generated by the latest schema. So we need to check the schema version of the all input rowsets and pick the latest tablet schema.

## What I'm doing:
Check the schema version of all input rowsets and pick the latest tablet schema as output rowset tablet schema.


Signed-off-by: sevev <qiangzh95@gmail.com>