[Enhancement] set required distribution flag in optExpression (#50140)
Why I'm doing:
update the existRequiredDistribution value in optExpression. The next rules need it to determine if we can change the distribution to adjust the plan because of skew data, bad statistics or something else.

What I'm doing:
If the plan had been optimized by the global dict, we can't change the distribution in the plan.
If the parent node had no requirement from its children, we can change the children's distribution.

Signed-off-by: packy92 <wangchao@starrocks.com>
Signed-off-by: zhiminr.ren <1240388654@qq.com>