[Doc] Ref docs (#50111)
> Note:
>
> - [ ] Merge this PR
> - [ ] Backport to 3.3, 3.2, 3.1
> - [ ] then, merge https://github.com/StarRocks/docs-site/pull/213
> - [ ] then, build staging
> - [ ] then build prod
> - [ ] then have Algolia crawl
> - [ ] then have Google crawl (submit sitemap.xml)

Our docs require too much micro-management of the navigation because we have all of the hundreds of reference pages in a very small number of folders. By creating a dir structure that matches the nav, we can let the static site generator (Docusaurus) auto generate the nav. This means that when we drop a file into a dir it gets added to the nav.

For example, if you look at the live site today and search for the word `SYNC` you will see that we have a reference page for the SQL statement SYNC. Unfortunately, it is not in the nav, and this causes the reader to wonder where they are in the docs. This is because we forgot to add an entry in the sidebars.js for this file.

Signed-off-by: DanRoscigno <dan@roscigno.com>
Signed-off-by: Dan Roscigno <dan@roscigno.com>