[BugFix] Make create_directories compatible with symbolic links (#50858)
create_directories() should be compatible with the scenario where dirname exists as a symbolic link
and linked to an existing directory. On CentOS create_directories() will fail in this situation,
but on Ubuntu, it won't. So we make a precheck here in order to have the same expected behavior
on both and probably all the other platforms.

Signed-off-by: Dejun Xia <xiadejun@starrocks.com>
(cherry picked from commit 3271d9055d36e8eceab4fe2e6c8f399103e2c3e2)