[Enhancement] A build machine that supports the AVX2 instruction set compiles a package without the AVX2 instruction set. (#51673)
Why I'm doing:
In some virtual machines or containers, the AVX2 instruction set is not supported. However, whether the compiled package from the current build script includes AVX2 support depends on the CPU instruction set of the build machine. Therefore, if the build machine supports the AVX2 instruction set, it is not possible to compile a package that does not include AVX2 support.

What I'm doing:
In the build.sh, add the --without-avx2 option to force the packaging of a product without the AVX2 instruction set. If this option is not added, the default strategy (detecting based on the build machine’s CPU) will be maintained.

Signed-off-by: happut <happut@163.com>