[BugFix] Fix analyzer exception for generated column rewrite
Why I'm doing:
This problem is introduce by pr ##50398. When we re-analyze the generated expression
in current scope for subquery, sometime we may meet an exception. In current implementation,
a map will be clear but the iteration does not break cause ConcurrentModificationException.

What I'm doing:
Remove the expression if re-analyze throw an exception instead of clear all generated column
expression.

Signed-off-by: srlch <linzichao@starrocks.com>