[BugFix] properly handle 100-continue in load action (#52582)
set connection close after issue the HTTP 307 redirection
when client sending http header 'Expect: 100-continue', it is allowed that the client still sends out the data even though the server doesn't respond a '100-Continue'. It will be hard for the server to know if the client sends the request body or not. The safest way is to close the connection no matter the client sends the request body or not.

Signed-off-by: Kevin Xiaohua Cai <caixiaohua@starrocks.com>
Signed-off-by: zhiminr.ren <1240388654@qq.com>