[Enhancement] Opt memory tracker for FE (#53055)
## Why I'm doing:
`SizeEstimator.estimate` will calculate the retained size of the object, that is the size of the param object and all the referenced objects. This will lead to repeated counting when the same object is referenced by many objects, and result in very high resource usage when the param object refers to network object or GlobalMetaStore.

## What I'm doing:
Change to `ClassLayout` to only calculate the shadow size of the object.

Signed-off-by: gengjun-git <gengjun@starrocks.com>
(cherry picked from commit 4f60313baa6c3cf1365996dd82a0b17a86585ab0)