[Doc] Fixed SHOW DATA command syntax in documentation. (#53498)
Why I'm doing:
The documentation seemed to show that the table name is optional even when the FROM clause was used. I think that's an error. When the FROM clause is used, I think the table name is required. The database name is optional.

What I'm doing:
Changing the line that shows the syntax of the SHOW DATA command.
Its a minor change.

Signed-off-by: kalyanmysore <kalyanmysore@yahoo.com>