[Enhancement] make LIKE operations' statistics consistent
The `LIKE(column, pattern)` function and the `column LIKE pattern`
operator have different statistics in
[ExpressionStatisticCalculator](https://github.com/StarRocks/starrocks/blob/41843adc63d8bb9c968370dc4d97acfb1b1a0685/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/statistics/ExpressionStatisticCalculator.java).

This leads to an issue when it's used together with other functions (e.g. `IF`).

Example:
`IF(column LIKE pattern, 'Yes', 'No')` estimates 2 distinct values as
expected, however when the same query is executed again as
`IF(LIKE(column, pattern), 'Yes', 'No')` the distinct value estimation
is now different than 2 because of this check [[1]].

Since they are the same operation, their statistics should be same too.
The `LikePredicateOperator` does not have a direct estimation in
`ExpressionStatisticCalculator`, instead it takes the statistics of the
first child by default [[2]].
With this change, the same applies to the `LIKE` function and it takes
the statistics of the first child (left hand side).

[1]: https://github.com/StarRocks/starrocks/blob/41843adc63d8bb9c968370dc4d97acfb1b1a0685/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/statistics/ExpressionStatisticCalculator.java#L189
[2]: https://github.com/StarRocks/starrocks/blob/41843adc63d8bb9c968370dc4d97acfb1b1a0685/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/statistics/ExpressionStatisticCalculator.java#L74

Signed-off-by: Elmi Ahmadov <e.ahmadov@celonis.com>