[Enhancement] Optimize Tablet Report (#54848)
## Why I'm doing:
In StarRocks, FE will periodically diff the tablets in BE and the tablets recorded in metadata, and then process the inconsistent tablets. The current implementation is that BE reports the full number of tablets to the FE Leader regularly (default 1 minute), and the Leader maintains a reporting queue, and then retrieves one BE tablet from the queue each time for single-threaded processing. For large-scale clusters, the speed of FE processing usually cannot keep up with the speed of BE reporting, resulting in the existence of all BE tablets in the reporting queue, which causes memory waste. This optimization uses the Leader's active pull mode to control the tablets in the reporting queue within a BE range.

## What I'm doing:
![image](https://github.com/user-attachments/assets/bfab21dc-9740-4e13-aa8c-a483a867d341)

After optimization, a new TabletController daemon is added to regularly pull the full number of tablets from the Backend. The pull condition is
1. For a certain BE, it has been more than collect_tablet_inverval_seconds since the last pull.
2. The processing queue of ReportHandler is empty.

BE still retains the ability to actively report tablets to FE Leader, but only for emergency situations, such as disk corruption and the need to immediately remove replicas from FE metadata.

## Test(a cluster with 5 million tablets)
after optimization
![image](https://github.com/user-attachments/assets/1eea3c3e-a430-4711-8b8c-62695222ad6d)

before optimization
![image](https://github.com/user-attachments/assets/d658afbd-74c9-40ac-8274-021d508f348c)

We can see the GC time has become smoother.

Signed-off-by: gengjun-git <gengjun@starrocks.com>