[Tool] Fix main build
The `main` branch build fails because of `glog` as follows:
```sh
thirdparty/src/glog-0.7.1
make: *** No targets specified and no makefile found. Stop.
```

The issue is `CMAKE_GENERATOR` is set to `Ninja` but it tries to compile
with `make`. To fix it, use the `BUILD_SYSTEM` command to compile which
is compatible with `CMAKE_GENERATOR`.

FIX #53383

Signed-off-by: Elmi Ahmadov <e.ahmadov@celonis.com>