[Enhancement] seperate flag for apply task for submit and running for primary key table
## Why I'm doing:
In current impl, tablet from primary key table will set _apply_running if the task has been submit into
apply thread pool. If the all threads are used up, the tasks will stuck in queue of the thread pool and all
caller of _stop_and_wait_apply_done will also stuck in this case. But this is not make sense because task has
not started yet. _stop_and_wait_apply_done should not be stuck in this case.

## What I'm doing:
seperate lock for apply task for submit and running. If the task has been submited, the submit flag will be set but
not running flag. The running flag will be set when the task has started. And the _stop_and_wait_apply_done
will wait if the running flag is set.

Signed-off-by: srlch <linzichao@starrocks.com>