[Refactor] Remove charge_mode from LRUCache (#55911)
Why I'm doing:
charge_size is the actual memory size occupied by an element, while value_size is the size of the element. For the underlying cache, it only needs to know charge_size for memory accounting purposes. As for the passed-in Item, if it needs to rely on the value_size for decoding, then ensuring that the data structure is self-descriptive.

The interface of DataCache only have one param charge_size, so In order to unify PageCache and DataCache, the charge mode of LRUCache will be removed.

What I'm doing:
Remove charge_mode from LRUCache

Signed-off-by: trueeyu <lxhhust350@qq.com>