[Enhancement] Optimize shard lock acquisition to reduce the holding time
## Why I'm doing:
In current impl, tablet meta lock or some primary key relative locks will be try to acquired after the shard lock has been holded in some place.
This may cause the shard lock to be held for a long time and cannot be released.

## What I'm doing:
Try to release the shard lock before try to acquire another lock if it is possible.

Signed-off-by: srlch <linzichao@starrocks.com>