[Doc] Japanese translation (#55725)
Add Japanese support for docs

@EsoragotoSpirit can you please work on the dictionary:

docs/translation/en-to-ja-configs/language_dicts/ja.yaml

I copied the English to Chinese one. The upstream project has Japanese support, and so they have a dictionary already for their terms. It may not be useful to you as they are in a different product space.

There are two translated files in this PR, I believe that they have a mix of Japanese and Chinese in them because I used the Chinese dictionary file. The README is updated with instructions to translate to Japanese. Also note the known issues in the README, I will work on those.

Signed-off-by: DanRoscigno <dan@roscigno.com>
Co-authored-by: EsoragotoSpirit <wanglichen@starrocks.com>
(cherry picked from commit 776379ed039cc31376fa5c541e12178c3ed469fb)