[BugFix] Reset index information when creating tablet for rollup job tablet
## Why I'm doing:
When creating tablet in rollup job, the tablet has index information which is the same as the base table
even if the rollup schema does not contain the index column. This will cause link inverted index files
failed because no ivt files will be generated.

## What I'm doing:
Reset index information based on the rollup schema when creating tablet in rollup job.

Signed-off-by: srlch <linzichao@starrocks.com>