[BugFix] Fix BE crash if auto increment column slot id is different from offset index
## Why I'm doing:
This problem is introduced by pr #56203. In #56203, we try to fix the auto increment column slot id.
But _auto_increment_slot_id was used for indexing the _output_tuple_desc->slots() before #56203. Because
before the fix, _auto_increment_slot_id means the offset index for the slot in _output_tuple_desc. But
after fix _auto_increment_slot_id as the slot id, it does not represent the position of the slot in
_output_tuple_desc which will cause a random crash in BE side when we try to access the invalid memory.

## What I'm doing:
Use get_slot_by_id to get the slot from _output_tuple_desc.

Signed-off-by: srlch <linzichao@starrocks.com>