[BugFix] Fix data unconsistent if add a new generated column which ref a normal column which contains cols file for pk table
## Why I'm doing:
In current impl, when we adding generated column, we get segment iterator as a non-primary key segment which
means that we do not want to apply the del vector on it to generate a cols file with the same row size as
the segment file. But the problem is, the flag is_primary_keys in segment read option is always false and
no dcg will be load from rocksdb. If the new added generated column has ref a column which has been updated
in partial update column mode with cols files, the result of the generated column will lost the updated value.

## What I'm doing:
Ensure the dcg will be load when adding generated column for pk table.

Signed-off-by: srlch <linzichao@starrocks.com>