[BugFix] Fix several bugs when restore table with random distribution
Bug 1:
If we restore a table with random distribution which contains multiple physical partitions in a single
logical partition, all physical partitions meta data will lost after FE restart exception the first one.
This is because we just add the tablet meta into global manager for the first physical partition when
we replay the editlog.

Bug 2:
If we restore a new partition into a existed table with random distribution, RestoreJob::resetPartitionForRestore
will be used to reset some ids for the new coming partition. It does not reset the default physical partition id
and even try to remove subpartition when we try to iterate it and get ConcurrentModificationException.

Bug 3:
After refactor for physical partition, ids for logical partition and physical partition are completely different.
We need to check parentId instead of id for physical partition in OlapTable::resetIdsForRestore

Solution:
For bug 1:
Make sure all physical partitions will be handled in RestoreJob::modifyInvertedIndex

For bug 2:
Reset ids in a suitable way

For bug 3:
Use parent id instead

Signed-off-by: srlch <linzichao@starrocks.com>