[Enhancement] Refactoring to extract authentication logical from Mysql negotiate (#57373)
Why I'm doing:
The current MySQL negotiate and authenticate logics are coupled together, which is not conducive to the expansion of subsequent PRs. Only after the user is authenticated can the connectContext be registered. However, some authentication modes require the use of connectContext, which makes development difficult.
On the other hand, it is not make sense to authenticate first and then register. Because after the connection is established, the ConnectionContext should be fully established, regardless of whether the authentication is successful.

What I'm doing:
extract authentication logical from Mysql negotiate
Modified process: 1. negotiate 2. register connection 3. authenticate