[BugFix] Fix auto increment partial update triggered by error if auto increment column is key when it miss in target columns in insert stmt
Why I'm doing:
If a auto increment column is defined as the key column in pk table and it miss in target columns in insert stmt, it will be triggered
partial update by error. In this case, partial update flag for auto increment will be set and get analyze exception.

What I'm doing:
Fix it, do not trigger if column is key.

Signed-off-by: srlch <linzichao@starrocks.com>