[Feature] Support role mapping with native user (#35719)
* Support role mapping with native user

* add more ut

Why I'm doing:
Currently, we support the Security Integration + Group Provider model, which is mainly used in the authentication phase. However, there is a problem with this model, which is that it must cooperate with an external authorization system, such as Ranger. This is mainly because the User generated by Security Integration is an external User and cannot cooperate with our current internal RBAC.

What I'm doing:
So we add a new syntax, grant role to external group. This will associate the permissions on our role with the user group.
Fixes #issue